+++ Examples/MMRecordPerformance/MMRecordAppDotNet/ADNPostsViewController.h
@@ -1,0 +1,13 @@
+//
+//  ADNPostsViewController.h
+//  MMRecordAppDotNet
+//
+//  Created by Conrad Stoll on 11/20/12.
+//  Copyright (c) 2012 Mutual Mobile. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface ADNPostsViewController : UITableViewController
+
+@end