+++ Examples/MMRecordPerformance/MMRecordAppDotNet/ADNUsersViewController.h
@@ -1,0 +1,13 @@
+//
+//  ADNUsersViewController.h
+//  MMRecordAppDotNet
+//
+//  Created by Conrad Stoll on 3/23/13.
+//  Copyright (c) 2013 Mutual Mobile. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+@interface ADNUsersViewController : UIViewController
+
+@end