+++ Examples/MMRecordPerformance/MMRecordAppDotNet/MMJSONPerformanceTestingServer.h
@@ -1,0 +1,15 @@
+//
+//  MMJSONPerformanceTestingServer.h
+//  MMRecordPerformance
+//
+//  Created by Conrad Stoll on 6/12/13.
+//  Copyright (c) 2013 Mutual Mobile. All rights reserved.
+//
+
+#import "MMJSONServer.h"
+
+@interface MMJSONPerformanceTestingServer : MMJSONServer
+
++ (void)setResultSetSize:(NSUInteger)resultSetSize;
+
+@end