+++ Examples/MMRecordPerformance/MMRecordAppDotNet/Source.m
@@ -1,0 +1,17 @@
+//
+//  Source.m
+//  MMRecordAppDotNet
+//
+//  Created by Conrad Stoll on 1/12/13.
+//  Copyright (c) 2013 Mutual Mobile. All rights reserved.
+//
+
+#import "Source.h"
+
+@implementation Source
+
+@dynamic clientID;
+@dynamic name;
+@dynamic link;
+
+@end